update go.mod.tpl for modern values (#18)
We had an old Go version (1.14) and ancient deps listed in the default go.mod.tpl file. This updates to a more modern set of deps and Go version (1.19) which results in fewer iterations of `go mod tidy` when generating a bootstrapped controller.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.